// app/src/generated/models/Uw_submissionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export const Uw_submissionsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Uw_submissionsstatecode = keyof typeof Uw_submissionsstatecode;
export const Uw_submissionsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Uw_submissionsstatuscode = keyof typeof Uw_submissionsstatuscode;
export const Uw_submissionsuw_complexity = {
  1: 'Simple',
  2: 'Standard',
  3: 'Complex'
} as const;
export type Uw_submissionsuw_complexity = keyof typeof Uw_submissionsuw_complexity;
export const Uw_submissionsuw_submissionstatus = {
  1: 'NewClaim_FNOL_',
  2: 'SupplementaryClaim',
  3: 'ReserveAdjustment',
  4: 'CoverageInquiry',
  5: 'ReopenedClaim'
} as const;
export type Uw_submissionsuw_submissionstatus = keyof typeof Uw_submissionsuw_submissionstatus;
export const Uw_submissionsuw_triagestatus = {
  1: 'New',
  2: 'InReview',
  3: 'Accepted',
  4: 'Reassigned',
  5: 'Reserved',
  6: 'Closed'
} as const;
export type Uw_submissionsuw_triagestatus = keyof typeof Uw_submissionsuw_triagestatus;

export interface Uw_submissions {
  createdby?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfby?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  importsequencenumber?: string;
  modifiedby?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfby?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunit?: string;
  owningbusinessunitname: string;
  owningteam?: string;
  owninguser?: string;
  statecode: Uw_submissionsstatecode;
  statecodename?: string;
  statuscode?: Uw_submissionsstatuscode;
  statuscodename?: string;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
  uw_acsquote?: string;
  uw_assignedunderwriter?: string;
  uw_assignmentreasoning?: string;
  uw_caseid: string;
  uw_classification?: string;
  uw_complexity?: Uw_submissionsuw_complexity;
  uw_complexityname?: string;
  uw_emailbody?: string;
  uw_emailreceived?: string;
  uw_emailsubject?: string;
  uw_executivesummary?: string;
  uw_insuredname?: string;
  uw_language?: string;
  uw_missingdocscount?: string;
  uw_missingdocuments?: string;
  uw_naicsprimarycode?: string;
  uw_processedon?: string;
  uw_producttype?: string;
  uw_region?: string;
  uw_reportfilename?: string;
  uw_reporttext?: string;
  uw_reporturl?: string;
  uw_senderemail?: string;
  uw_sendername?: string;
  uw_submissionid: string;
  uw_submissionstatus?: Uw_submissionsuw_submissionstatus;
  uw_submissionstatusname?: string;
  uw_triagestatus?: Uw_submissionsuw_triagestatus;
  uw_triagestatusname?: string;
  uw_underwriteremail?: string;
  versionnumber?: string;
}