// app/src/generated/models/Uw_submissiondocumentsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export const Uw_submissiondocumentsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Uw_submissiondocumentsstatecode = keyof typeof Uw_submissiondocumentsstatecode;
export const Uw_submissiondocumentsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Uw_submissiondocumentsstatuscode = keyof typeof Uw_submissiondocumentsstatuscode;
export const Uw_submissiondocumentsuw_status = {
  1: 'Provided',
  2: 'Missing',
  3: 'Unclear'
} as const;
export type Uw_submissiondocumentsuw_status = keyof typeof Uw_submissiondocumentsuw_status;

export interface Uw_submissiondocuments {
  createdby?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfby?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  importsequencenumber?: string;
  modifiedby?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfby?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunit?: string;
  owningbusinessunitname: string;
  owningteam?: string;
  owninguser?: string;
  statecode: Uw_submissiondocumentsstatecode;
  statecodename?: string;
  statuscode?: Uw_submissiondocumentsstatuscode;
  statuscodename?: string;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
  uw_documenttype?: string;
  uw_name: string;
  uw_sharepointurl?: string;
  uw_status?: Uw_submissiondocumentsuw_status;
  uw_statusname?: string;
  uw_submissiondocumentid: string;
  uw_submissionid: string;
  uw_submissionidname?: string;
  versionnumber?: string;
}