// app/src/generated/models/Uw_classificationmatchsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export const Uw_classificationmatchsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Uw_classificationmatchsstatecode = keyof typeof Uw_classificationmatchsstatecode;
export const Uw_classificationmatchsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Uw_classificationmatchsstatuscode = keyof typeof Uw_classificationmatchsstatuscode;
export const Uw_classificationmatchsuw_confidence = {
  1: 'High',
  2: 'Medium',
  3: 'Low'
} as const;
export type Uw_classificationmatchsuw_confidence = keyof typeof Uw_classificationmatchsuw_confidence;
export const Uw_classificationmatchsuw_isprimary = {
  0: 'No',
  1: 'Yes'
} as const;
export type Uw_classificationmatchsuw_isprimary = keyof typeof Uw_classificationmatchsuw_isprimary;

export interface Uw_classificationmatchs {
  createdby?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfby?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  importsequencenumber?: string;
  modifiedby?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfby?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunit?: string;
  owningbusinessunitname: string;
  owningteam?: string;
  owninguser?: string;
  statecode: Uw_classificationmatchsstatecode;
  statecodename?: string;
  statuscode?: Uw_classificationmatchsstatuscode;
  statuscodename?: string;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
  uw_classificationmatchid: string;
  uw_confidence?: Uw_classificationmatchsuw_confidence;
  uw_confidencename?: string;
  uw_isprimary?: Uw_classificationmatchsuw_isprimary;
  uw_isprimaryname?: string;
  uw_naicscode: string;
  uw_reasoning?: string;
  uw_submissionid: string;
  uw_submissionidname?: string;
  uw_title?: string;
  versionnumber?: string;
}